requirements.txt
msgraph-sdk>=1.0.0
azure-identity>=1.15.0
python-dotenv>=1.0.0
prompt-toolkit>=3.0.0
rich>=13.0.0
